+++ src/runtime/typemanager.cs
@@ -4,10 +4,10 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using Python.Runtime.Platform;
+using Python.Runtime.Slots;
 
 namespace Python.Runtime
 {
-
     /// <summary>
     /// The TypeManager class is responsible for building binary-compatible
     /// Python type objects that are implemented in managed code.
@@ -154,6 +154,14 @@ namespace Python.Runtime
             Marshal.WriteIntPtr(type, TypeOffset.tp_dictoffset, (IntPtr)tp_dictoffset);
 
             InitializeSlots(type, impl.GetType());
+
+            if (typeof(IGetAttr).IsAssignableFrom(clrType)) {
+                InitializeSlot(type, TypeOffset.tp_getattro, typeof(SlotOverrides).GetMethod(nameof(SlotOverrides.tp_getattro)));
+            }
+
+            if (typeof(ISetAttr).IsAssignableFrom(clrType)) {
+                InitializeSlot(type, TypeOffset.tp_setattro, typeof(SlotOverrides).GetMethod(nameof(SlotOverrides.tp_setattro)));
+            }
 
             if (base_ != IntPtr.Zero)
             {
@@ -764,6 +772,11 @@ namespace Python.Runtime
             var offset = (int)fi.GetValue(typeOffset);
 
             Marshal.WriteIntPtr(type, offset, slot);
+        }
+
+        static void InitializeSlot(IntPtr type, int slotOffset, MethodInfo method) {
+            IntPtr thunk = Interop.GetThunk(method);
+            Marshal.WriteIntPtr(type, slotOffset, thunk);
         }
 
         /// <summary>